requirements.txt
langchain
langchain-community
langchain-openai
langgraph
openai
chromadb
sqlalchemy
streamlit
pypdf
python-dotenv
fastmcp